+++ common/RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## in develop
 
 * Adds optional `localizationDir` parameter to `LocalizationDisambiguator` methods for specifying the localization directory that must be used
+* Fixes `prettyFormat` function to handle case clases with public fields in the second parameter list
 
 ## 0.4.1 (2021-06-08)