+++ common/RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # dxCommon
 
-## in develop
+## ## 0.5.0 (2021-06-23)
 
 * Adds optional `localizationDir` parameter to `LocalizationDisambiguator` methods for specifying the localization directory that must be used
 * Fixes `prettyFormat` function to handle case clases with public fields in the second parameter list